Add Apple Silicon (MPS) support for depth-aware editing on Mac.
Port CUDA-only paths to MPS/CPU, add Mac setup docs and pinned deps, and fix weight downloads and PyTorch 2.8 checkpoint loading for local inference.

+++ src/featglac/model/transformer_2d.py
@@ -409,7 +409,7 @@ class Transformer2DModel(ModelMixin, ConfigMixin):
             logits = logits.permute(0, 2, 1)
 
             # log(p(x_0))
-            output = F.log_softmax(logits.double(), dim=1).float()
+            output = F.log_softmax(logits.float(), dim=1)
 
         if self.is_input_patches:
             if self.config.norm_type != "ada_norm_single":
